GWJ/4_streamlit/requirements.txt
# s3fs
geopandas
plotly
folium
streamlit_folium

# requirements.txt
# s3fs==x.x.x
s3fs
# Direct pypi install coming soon
git+https://github.com/streamlit/files-connection